\name{testCheckboxes}
\alias{testCheckboxes}
\title{Test that checkboxes get the appropriate value}
\usage{
  testCheckboxes(checkbox1 = GalaxyLogicalParam(),
    checkbox2 = GalaxyLogicalParam(),
    outfile = GalaxyOutput("myoutput", "txt"))
}
\arguments{
  \item{checkbox1}{A checkbox to be set to TRUE}

  \item{checkbox2}{A checkbox to be set to FALSE}

  \item{optionalParam}{An optional parameter}

  \item{outfile}{where to write diagnostic output}
}
\description{
  This function displays the values of its parameters, or
  'missing' if a parameter value is not supplied.
}
\details{
  These are the details.
}

