\name{testRequiredOption}
\alias{testRequiredOption}
\title{Test whether a required parameter works as expected.}
\usage{
  testRequiredOption(requiredOption = GalaxyCharacterParam(required = TRUE, requiredMsg = "THIS FIELD IS MANDATORY"),
    outputfile = GalaxyOutput("mytext", "txt"))
}
\arguments{
  \item{requiredOption}{An option that is required.}

  \item{outputfile}{The name of a file where diagnostic
  output is written.}
}
\description{
  This function tests whether a required function works as
  expected.
}
\details{
  These are the details.
}

